+++ app/javascript/src/config/action_types.js
@@ -3,10 +3,6 @@ export const SET_CURRENT_TEAM = "SET_CURRENT_TEAM";
 export const GET_LIST_OF_TEAMS = "GET_LIST_OF_TEAMS";
 export const SET_TEAM_DETAILS = "SET_TEAM_DETAILS";
 
-// activities
-export const GLOBAL_ACTIVITIES_DATA = "GLOBAL_ACTIVITIES_DATA";
-export const DESTROY_GLOBAL_ACTIVITIES_DATA = "DESTROY_GLOBAL_ACTIVITIES_DATA";
-
 // users
 export const SET_CURRENT_USER = "SET_CURRENT_USER";
 export const CHANGE_CURRENT_USER_FULL_NAME = "CHANGE_CURRENT_USER_FULL_NAME";
