+++ app/javascript/src/config/api_endpoints.js
@@ -1,5 +1,3 @@
-// activities
-export const ACTIVITIES_PATH = "/client_api/activities";
 
 // settings
 export const SETTINGS_PATH = "/settings";
