+++ app/javascript/src/services/api/endpoints.js
@@ -1,0 +1,2 @@
+// activities
+export const ACTIVITIES_PATH = "/client_api/activities";
