+++ config/routes/activities.rb
@@ -1,0 +1,2 @@
+# activities
+get '/activities', to: 'activities#index'
