+++ config/routes/notifications.rb
@@ -1,0 +1,2 @@
+# notifications
+get '/recent_notifications', to: 'notifications#recent_notifications'
