+++ package.json
@@ -29,6 +29,7 @@
     "eslint-plugin-jsx-a11y": "^6.0.2",
     "eslint-plugin-prettier": "^2.1.2",
     "eslint-plugin-react": "^7.1.0",
+    "flow-bin": "^0.56.0",
     "prettier": "^1.7.0",
     "webpack-dev-server": "^2.5.1"
   },
@@ -68,8 +69,6 @@
     "react-bootstrap-table": "^4.0.0",
     "react-bootstrap-timezone-picker": "^1.0.11",
     "react-data-grid": "^2.0.2",
-    "react-tagsinput": "^3.17.0",
-    "react-transition-group": "^2.2.0",
     "react-dom": "^15.6.1",
     "react-intl": "^2.3.0",
     "react-intl-redux": "^0.6.0",
@@ -78,7 +77,9 @@
     "react-router-bootstrap": "^0.24.2",
     "react-router-dom": "^4.1.2",
     "react-router-prop-types": "^0.0.1",
+    "react-tagsinput": "^3.17.0",
     "react-timezone": "^0.2.0",
+    "react-transition-group": "^2.2.0",
     "redux": "^3.7.2",
     "redux-thunk": "^2.2.0",
     "resolve-url-loader": "^2.1.0",
